refactor(client): move voice and video into its own settings page
